requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
pandas==2.1.4
numpy==1.24.3
tensorflow==2.15.0
prophet==1.1.5
cvxpy==1.4.1
pulp==2.7.0
scikit-learn==1.3.2
httpx==0.25.2
asyncio-mqtt==0.13.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aioredis==2.0.1
celery==5.3.4
plotly==5.17.0
streamlit==1.28.2
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
